test_amt.h: Add missing function prototype that causes test segfault
ferite_amt-hash-set-and-get.c segfaults when run with libgc-ferite due
to return value type mismatch for the function
_ferite_hamt_get_with_key() between the one set in ferite_amt.c
(FeriteAMTNode *) and the one implicitly set by the compiler (int) due
to the lack of prototype - the returned pointer overflows when casted to
int which results in invalid memory address hence the segfault.

Also add the prototype for ferite_hamt_get as well. While we have not
seen any test segfault due to it, theoretically it is possible.